refactor: organize `languages` namespace
